Condition
Write a program that takes an integer from the console and prints the corresponding month. If the number is more than 12 or less than 1 print "Error!".
Input:
You will receive a single integer on a single line.
Output:
If the number is within the boundaries, print the corresponding month, otherwise, print "Error!".