Condition
Write a program that prints the language that a given country speaks. You can receive only the following combinations: English is spoken in England and USA; Spanish is spoken in Spain, Argentina, and Mexico; for the others, we should print "unknown".
Input:
You will receive a single country name on a single line.
Output:
Print the language, which the country speaks, or if it is unknown for your program, print "unknown".