Condition
You will be given a string representing a username. The password will be that username reversed. Until you receive the correct password, print on the console "Incorrect password. Try again.". When you receive the correct password, print "User {username} logged in." However, on the fourth try, if the password is still not correct, print "User {username} blocked!" and end the program.