Orders
Write a method that calculates the total price of an order and prints it on the console. The method should receive one of the following products: coffee, water, coke, snacks, and a quantity of the product. The prices for a single piece of each product are:
•	coffee – 1.50
•	water – 1.00
•	coke – 1.40
•	snacks – 2.00
Print the result rounded to the second decimal place.
Example:
Input:        Output:
water         5.00
5